Add documentation on version 0.6 to the main page
When people come to the library main page from the RobotFramework
project page, or Google, they expect to see information about the latest
version of the library. But this page shows v0.5 as the latest, although
the latest package has version 0.6.

Here are docs I've generated using RobotFramework LibDoc tool upon the
0.6 version of the library (upon the commit marked with tag 0.6).