If statement in disconnect_from_database
When using 'Disconnect from Database' in a Suite Teardown, while no connection was created because of an early error in the TC. The shown message is not really tidy; "AttributeError: 'NoneType' object has no attribute 'close'" and the Teardown Fails. 
Therefore I created a simple If Else statement that if no connection is open this will be returned as a message and the Teardown wil not fail.