Added fallback conditions to ConfigReader.get, **kwargs to ConnectionManager
Some database connection functions don't require all of dbapiModuleName, dbName, dbUsername...

These functions also fail in some libraries if given a blank string on some libraries. 

By 'some libraries' here I mostly mean pymssql, but these changes should be invisible to any tests running with a fully db.cfg. 

Similarly, I passed **kwargs in from connect_to_database on connections to misc databases not in the existing if/elif cases. This was mostly to allow for passing in tds_version on pymssql (see this issue: https://github.com/pymssql/pymssql/issues/860), but it would be useful for any misc database to be able to set up their connection using custom parameters. 

It may also be a good idea to pass kwargs in to each db_connection in that whole big if/else statement, but I wanted this to be minimally impactful to connection modules I don't know well enough to not break.